Refactor: Move interfaces to separate model files for better organization
