+++ frontend/src/app/models/create-session-request.model.ts
@@ -1,0 +1,3 @@
+export interface CreateSessionRequest {
+  name: string;
+}
