+++ frontend/src/app/models/index.ts
@@ -1,0 +1,26 @@
+// Question and Quiz related interfaces
+export * from './question.model';
+export * from './quiz-session.model';
+export * from './quiz-state.model';
+export * from './sample-question.model';
+
+// Team and User interfaces
+export * from './team.model';
+export * from './user-session.model';
+
+// Answer related interfaces
+export * from './answer.model';
+
+// Request interfaces
+export * from './create-question-request.model';
+export * from './create-session-request.model';
+export * from './join-session-request.model';
+export * from './submit-answer-request.model';
+export * from './score-answer-request.model';
+export * from './load-sample-questions-request.model';
+
+// Socket related interfaces
+export * from './join-session-data.model';
+export * from './submit-answer-data.model';
+export * from './presenter-action.model';
+export * from './socket-events.model';
