+++ frontend/src/app/models/join-session-data.model.ts
@@ -1,0 +1,4 @@
+export interface JoinSessionData {
+  sessionCode: string;
+  teamName: string;
+}
