+++ frontend/src/app/models/join-session-request.model.ts
@@ -1,0 +1,4 @@
+export interface JoinSessionRequest {
+  sessionCode: string;
+  teamName: string;
+}
