+++ frontend/src/app/models/load-sample-questions-request.model.ts
@@ -1,0 +1,4 @@
+export interface LoadSampleQuestionsRequest {
+  sessionCode: string;
+  roundNumber: number;
+}
