+++ frontend/src/app/models/presenter-action.model.ts
@@ -1,0 +1,5 @@
+export interface PresenterAction {
+  sessionCode: string;
+  action: 'start_question' | 'end_question' | 'show_leaderboard' | 'show_review' | 'next_round';
+  questionId?: string;
+}
