+++ frontend/src/app/models/quiz-state.model.ts
@@ -1,0 +1,9 @@
+import { Question } from './question.model';
+
+export interface QuizState {
+  currentQuestion?: Question;
+  timeRemaining?: number;
+  isActive: boolean;
+  submissionsReceived: number;
+  totalTeams: number;
+}
