+++ frontend/src/app/models/score-answer-request.model.ts
@@ -1,0 +1,4 @@
+export interface ScoreAnswerRequest {
+  isCorrect: boolean;
+  pointsAwarded: number;
+}
