+++ frontend/src/app/models/submit-answer-data.model.ts
@@ -1,0 +1,6 @@
+export interface SubmitAnswerData {
+  sessionCode: string;
+  teamId: string;
+  questionId: string;
+  answer: string | string[];
+}
