+++ frontend/src/app/models/submit-answer-request.model.ts
@@ -1,0 +1,5 @@
+export interface SubmitAnswerRequest {
+  questionId: string;
+  teamId: string;
+  answer: string | string[];
+}
