+++ frontend/src/app/models/team.model.ts
@@ -1,0 +1,8 @@
+export interface Team {
+  id: string;
+  quiz_session_id: string;
+  name: string;
+  total_points: number;
+  joined_at: string;
+  last_activity: string;
+}
