+++ frontend/src/app/models/user-session.model.ts
@@ -1,0 +1,5 @@
+export interface UserSession {
+  teamId: string;
+  teamName: string;
+  sessionCode: string;
+}
